+++ src/log-store/src/kafka/util.rs
@@ -12,5 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod offset;
 pub mod record;
