+++ src/log-store/src/test_util.rs
@@ -12,6 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-pub mod kafka;
 pub mod log_store_util;
